% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_methods.R
\name{download_gbif}
\alias{download_gbif}
\title{Download the GBIF taxonomy data dump to a temporary directory}
\usage{
download_gbif()
}
\value{
A character vector containing paths to the relevant downloaded and unzipped GBIF data dump files.
}
\description{
Download the GBIF taxonomy data dump to a temporary directory
}
\examples{
\dontrun{download_gbif()}
}
