% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_methods.R
\name{fuzzy_search}
\alias{fuzzy_search}
\title{Match scientific names that might be misspelled or that contain partial information}
\usage{
fuzzy_search(x, term, sensitivity = 0, allow_term_removal = FALSE)
}
\arguments{
\item{x}{A tibble created with \code{load_taxonomies()} or \code{load_population()} or \code{load_sample()}.}

\item{term}{A string consisting of a scientific name}

\item{sensitivity}{Mismatch tolerance (defaults to intolerant i.e. 0)}

\item{allow_term_removal}{Allow search against only the first word of a search query. Useful
when "Genus sp." or "Genus indet." is the search phrase.}
}
\value{
A list of candidate match(es), if applicable.
}
\description{
Match scientific names that might be misspelled or that contain partial information
}
\examples{
fuzzy_search(load_sample(), "Miacis deutschi")
fuzzy_search(load_sample(), "Miacis sp.", allow_term_removal = TRUE)
fuzzy_search(load_sample(), "Miacus deutschi", sensitivity = 1)
}
