% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_methods.R
\name{term_conversion}
\alias{term_conversion}
\title{Convert GBIF terms to NCBI terms}
\usage{
term_conversion(x)
}
\arguments{
\item{x}{A tibble created with \code{load_taxonomies()} or \code{load_population()} or \code{load_sample()}.}
}
\value{
A tibble with converted terms.
}
\description{
\code{term_conversion} converts GBIF terminology to NCBI terminology where there is
no biological provenance for the difference.
}
\examples{
term_conversion(load_sample())
}
