% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_methods.R
\name{download_ncbi}
\alias{download_ncbi}
\title{Download the NCBI taxonomy data dump to a temporary directory}
\usage{
download_ncbi()
}
\value{
A character vector containing paths to the downloaded and unzipped NCBI data dump
}
\description{
Download the NCBI taxonomy data dump to a temporary directory
}
\examples{
\dontrun{download_ncbi()}
}
