% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_methods.R
\name{load_population}
\alias{load_population}
\title{Load previously merged GBIF and NCBI taxonomies}
\usage{
load_population(x)
}
\arguments{
\item{x}{Path to a text file containing previously merged NCBI and GBIF taxonomies (compressed or uncompressed).}
}
\value{
A tibble
}
\description{
\code{load_population()} imports a previously merged taxonomy from your file system. An
example of an previously merged taxonomy can be downloaded
from \url{https://drive.google.com/file/d/1gpvm9QKdOcuGo_cIXPkAgGlB-qfKZZU6/view?usp=sharing}.
}
\examples{
\dontrun{load_population("path/to/merged_taxonomies")}
}
