% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_mdb}
\alias{plot_mdb}
\title{Generic for plot_mdb methods}
\usage{
plot_mdb(x)
}
\arguments{
\item{x}{An object of the class one_rank or the class all_ranks.}
}
\value{
A \code{ggplot2} derived plot
}
\description{
Generic for plot_mdb methods
}
\details{
A generic with methods that plot \code{taxonbridge} data types (\code{one_rank} and \code{all_ranks}). These
data types are created by using the methods \code{prepare_rank_dist()} or \code{prepare_comparable_rank_dist()}.
}
\examples{
plot_mdb(prepare_rank_dist(load_sample(), NCBI = TRUE, GBIF = TRUE))
plot_mdb(prepare_comparable_rank_dist(load_sample()))
plot_mdb(prepare_rank_dist(get_status(load_sample(),status = "synonym"), NCBI = TRUE))
plot_mdb(prepare_comparable_rank_dist(get_validity(get_status(load_sample()), valid = TRUE)))
}
